The TemporalResolution was truncated to int.
When a data set containing TemporalResolution was read into VTK,
the value was truncated.  This caused incorrect time information
for fMRI.

+++ Source/vtkDICOMSliceSorter.cxx
@@ -469,7 +469,7 @@ void vtkDICOMSliceSorter::SortFiles(vtkIntArray *files, vtkIntArray *frames)
   }
 
   // important time-related variables
-  int temporalSpacing = 1.0;
+  double temporalSpacing = 1.0;
 
   if (meta->Has(DC::SharedFunctionalGroupsSequence))
   {
